removed index from sphere intersection, created world header and moved appropriate structs to new header, added header to makefile

+++ srcs/world.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   world.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: thakala <thakala@student.42.fr>            +#+  +:+       +#+        */
+/*   By: deelliot <deelliot@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/23 10:18:04 by thakala           #+#    #+#             */
-/*   Updated: 2022/10/25 17:43:16 by thakala          ###   ########.fr       */
+/*   Updated: 2022/10/26 12:40:01 by deelliot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,10 +87,10 @@ void	default_world(t_world *world)
 			default_material_1());
 	sphere_2 = sphere(default_origin(), default_transform_2(),
 			default_material_2());
-	if (vec_push(&world->objects, &sphere_1) == (uint64_t)VEC_ERROR)
+	if (vec_push(&world->objects, &sphere_1) == VEC_ERROR)
 		handle_errors("unable to malloc for world object");
-	if (vec_push(&world->objects, &sphere_2) == (uint64_t)VEC_ERROR)
+	if (vec_push(&world->objects, &sphere_2) == VEC_ERROR)
 		handle_errors("unable to malloc for world object");
-	if (vec_push(&world->lights, &light) == (uint64_t)VEC_ERROR)
+	if (vec_push(&world->lights, &light) == VEC_ERROR)
 		handle_errors("unable to malloc for light");
 }
